large fix some issues with it

+++ fastapi_helpers/crud/BaseCrud.py
@@ -1,7 +1,7 @@
 from ormar import Model
 from fastapi_helpers.routes.Paginate import paginate_object, load_data_callback, Pagination
-from fastapi import HTTPException
-from typing import Any, Type
+from typing import Any, List, Type
+
 
 class BaseCrud():
 
@@ -13,7 +13,7 @@ class BaseCrud():
     async def load_data(self, result=[]) -> Any:
         return await load_data_callback(result)
 
-    async def get_list(self, options: Pagination):
+    async def get_list(self, options: Pagination) -> List[Any]:
         if(options.search == ''):
             r = await paginate_object(
                 self.model,
@@ -23,16 +23,16 @@ class BaseCrud():
             return r
         else:
             return await self.search(options)
-    
+
     async def search(self, options: Pagination):
         searchable = {}
         for item in self.search_attrs:
             searchable[item + "__icontains"] = options.search
         options.orable = searchable
         return await paginate_object(
-                self.model,
-                options,
-                (self.load_data, {})
+            self.model,
+            options,
+            (self.load_data, {})
         )
 
     async def get(self, id: str) -> Any:
@@ -46,7 +46,7 @@ class BaseCrud():
         )
         if(len(objs) > 0):
             return objs[0]
-        raise HTTPException(status_code=404, detail="Item not found")
+        return None
 
     async def create(self, model_in) -> Any:
         params = to_dict(model_in)
@@ -56,14 +56,19 @@ class BaseCrud():
 
     async def update(self, id, model_in) -> Any:
         params = to_dict(model_in)
-        obj = await self.model.objects.get(id=id)
+        obj = await self.model.objects.get_or_none(id=id)
+        if(obj is None):
+            return None
         obj = await obj.update(**params)
         return obj
 
     async def delete(self, id) -> Any:
-        obj = await self.model.objects.get(id=id)
+        obj = await self.model.objects.get_or_none(id=id)
+        if(obj is None):
+            return None
         await obj.delete()
         return {'state': 'ok'}
+
 
 def to_dict(model_in):
     params = {}
